Extract settings utilities and update password translations
