+++ src/translations/pt-BR.json
@@ -108,11 +108,38 @@
       "shareThoughts": "Compartilhe seus pensamentos"
     }
   },
+  "settings": {
+    "title": "Configurações",
+    "description": "Gerencie as preferências da sua conta.",
+    "name": "Nome",
+    "phone": "Telefone",
+    "language": "Idioma",
+    "unsavedChanges": "Você tem alterações não salvas",
+    "apply": "Aplicar",
+    "interactionCancelled": "Interação cancelada",
+    "languageUpdated": "Idioma atualizado",
+    "updating": "Atualizando configurações...",
+    "updateSuccess": "Configurações atualizadas",
+    "updateError": "Falha ao atualizar as configurações"
+  },
+  "promptPassword": {
+    "title": "Confirmar senha",
+    "description": "Por favor, digite sua senha para continuar.",
+    "submit": "Confirmar"
+  },
   "soon": "Em breve",
   "validation": {
     "invalidEmail": "Email inválido",
     "passwordMin": "Mínimo de {{count}} caracteres",
     "passwordMismatch": "As senhas não coincidem",
-    "required": "Campo obrigatório"
+    "required": "Campo obrigatório",
+    "error": "Erro"
+  },
+  "appwrite": {
+    "errors": {
+      "user_invalid_credentials": {
+        "description": "Credenciais inválidas"
+      }
+    }
   }
 }
